Phonepe_Pulse_Data_Visualization/requirements.txt
anaconda-navigator==2.3.0
anaconda-project==0.10.2
astropy==5.0.4
autopep8==1.6.0
binarytree==6.5.1
bitarray==2.4.1
bkcharts==0.2
boto3==1.21.32
bottleneck==1.3.4
brotlipy==0.7.0
comtypes==1.1.10
conda-build==3.21.8
conda-package-handling==1.8.1
conda-repo-cli==1.0.4
conda-token==0.3.0
conda-verify==3.4.2
cython==0.29.28
cytoolz==0.11.0
datashader==0.13.0
distributed==2022.2.1
flake8==3.9.2
gensim==4.1.2
glob2==0.7
google-cloud-storage==1.31.0
gradio==3.18.0
grpcio==1.42.0
h5py==3.6.0
hvplot==0.7.3
imagecodecs==2021.8.26
imbalanced-learn==0.10.1
intake==0.6.5
jdcal==1.4.1
jupyter==1.0.0
jupyter-dash==0.4.2
jupyterlab==3.3.2
linkify-it-py==1.0.3
mdit-py-plugins==0.3.3
missingno==0.5.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mock==4.0.3
munkres==1.1.4
mypy-extensions==0.4.3
mysql-connector-python==8.0.32
navigator-updater==0.2.1
nltk==3.7
nose==1.3.7
olefile==0.46
openpyxl==3.0.9
pep8==1.7.1
pip-chill==1.0.1
pycurl==7.44.1
pydocstyle==6.1.1
pyhamcrest==2.0.2
pyjwt==2.1.0
pymongo==4.3.3
pyodbc==4.0.32
pyreadline==2.1
pysocks==1.7.1
pytest==7.1.1
python-snappy==0.6.0
rope==0.22.0
scikit-image==0.19.2
scikit-learn-intelex==2021.20220215.102710
scrapy==2.6.1
sip==4.19.13
sklearn==0.0.post1
snscrape==0.6.1.20230314
sortedcollections==2.1.0
spyder==5.1.5
sqlalchemy==1.4.32
statsmodels==0.13.2
streamlit-option-menu==0.3.2
sympy==1.10.1
tables==3.6.1
tabulate==0.8.9
tbb==0.2
unidecode==1.2.0
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
xlrd==2.0.1
xlsxwriter==3.0.3
xlwings==0.24.9
yapf==0.31.0
